Remove AB test for ad permission
