+++ src/shared/model/experiments/abSwitches.ts
@@ -1,6 +1,5 @@
 export const abSwitches = {
   abExampleTest: false,
-  abAdvertisingPermissionRegistrationPrompt: true,
   abOptInPromptPostSignInImage: true,
   abMarketingConsentImage: true,
 };
